Move InoreaderUrlReader into a separate file
