+++ include/inoreaderapi.h
@@ -35,22 +35,6 @@ private:
 	std::string auth_header;
 };
 
-class InoreaderUrlReader : public UrlReader {
-public:
-	InoreaderUrlReader(ConfigContainer* c,
-		const std::string& url_file,
-		RemoteApi* a);
-	virtual ~InoreaderUrlReader();
-	virtual void write_config();
-	virtual void reload();
-	virtual std::string get_source();
-
-private:
-	ConfigContainer* cfg;
-	std::string file;
-	RemoteApi* api;
-};
-
 } // namespace newsboat
 
 #endif /* NEWSBOAT_INOREADERAPI_H_ */
