+++ include/inoreaderurlreader.h
@@ -1,0 +1,29 @@
+#ifndef NEWSBOAT_INOREADERURLREADER_H_
+#define NEWSBOAT_INOREADERURLREADER_H_
+
+#include "urlreader.h"
+
+namespace newsboat {
+
+class ConfigContainer;
+class RemoteApi;
+
+class InoreaderUrlReader : public UrlReader {
+public:
+	InoreaderUrlReader(ConfigContainer* c,
+		const std::string& url_file,
+		RemoteApi* a);
+	virtual ~InoreaderUrlReader();
+	virtual void write_config();
+	virtual void reload();
+	virtual std::string get_source();
+
+private:
+	ConfigContainer* cfg;
+	std::string file;
+	RemoteApi* api;
+};
+
+} // namespace newsboat
+
+#endif /* NEWSBOAT_INOREADERURLREADER_H_ */
