+++ mk/mk.deps
@@ -94,14 +94,15 @@ src/controller.o: src/controller.cpp include/controller.h include/cache.h \
  include/configexception.h include/configparser.h include/dbexception.h \
  include/downloadthread.h include/exception.h include/feedhqapi.h \
  include/feedhqurlreader.h include/fileurlreader.h include/globals.h \
- include/inoreaderapi.h include/itemrenderer.h include/htmlrenderer.h \
- include/textformatter.h include/logger.h include/newsblurapi.h \
- rss/feed.h rss/item.h include/ocnewsapi.h include/oldreaderapi.h \
- include/opmlurlreader.h include/regexmanager.h include/remoteapi.h \
- include/rssparser.h include/stflpp.h include/strprintf.h \
- include/ttrssapi.h 3rd-party/json.hpp include/utils.h include/view.h \
- include/controller.h include/filebrowserformaction.h \
- include/formaction.h include/history.h include/keymap.h include/stflpp.h
+ include/inoreaderapi.h include/inoreaderurlreader.h \
+ include/itemrenderer.h include/htmlrenderer.h include/textformatter.h \
+ include/logger.h include/newsblurapi.h rss/feed.h rss/item.h \
+ include/ocnewsapi.h include/oldreaderapi.h include/opmlurlreader.h \
+ include/regexmanager.h include/remoteapi.h include/rssparser.h \
+ include/stflpp.h include/strprintf.h include/ttrssapi.h \
+ 3rd-party/json.hpp include/utils.h include/view.h include/controller.h \
+ include/filebrowserformaction.h include/formaction.h include/history.h \
+ include/keymap.h include/stflpp.h
 src/dialogsformaction.o: src/dialogsformaction.cpp \
  include/dialogsformaction.h include/formaction.h include/history.h \
  include/keymap.h include/configparser.h include/configactionhandler.h \
@@ -225,11 +226,11 @@ src/inoreaderapi.o: src/inoreaderapi.cpp include/inoreaderapi.h \
  include/strprintf.h include/remoteapi.h include/urlreader.h \
  include/strprintf.h include/utils.h
 src/inoreaderurlreader.o: src/inoreaderurlreader.cpp \
- include/inoreaderapi.h include/cache.h include/configcontainer.h \
- include/configparser.h include/configactionhandler.h include/rss.h \
- include/matcher.h filter/FilterParser.h include/utils.h include/logger.h \
- config.h include/strprintf.h include/remoteapi.h include/urlreader.h \
- include/fileurlreader.h include/logger.h
+ include/inoreaderurlreader.h include/urlreader.h \
+ include/configcontainer.h include/configparser.h \
+ include/configactionhandler.h include/fileurlreader.h include/logger.h \
+ config.h include/strprintf.h include/remoteapi.h \
+ include/configcontainer.h include/utils.h include/logger.h
 src/itemlistformaction.o: src/itemlistformaction.cpp \
  include/itemlistformaction.h include/history.h include/listformaction.h \
  include/formaction.h include/keymap.h include/configparser.h \
