+++ src/inoreaderurlreader.cpp
@@ -1,7 +1,10 @@
-#include "inoreaderapi.h"
+#include "inoreaderurlreader.h"
 
+#include "configcontainer.h"
 #include "fileurlreader.h"
 #include "logger.h"
+#include "remoteapi.h"
+#include "utils.h"
 
 namespace newsboat {
